Moved the method and used in the command runner
