+++ PHPCI/Command/RunCommand.php
@@ -15,6 +15,7 @@ use PHPCI\Helper\Lang;
 use PHPCI\Logging\BuildDBLogHandler;
 use PHPCI\Logging\LoggedBuildContextTidier;
 use PHPCI\Logging\OutputLogHandler;
+use PHPCI\Service\BuildService;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -97,14 +98,16 @@ class RunCommand extends Command
         $store = Factory::getStore('Build');
         $result = $store->getByStatus(0, $this->maxBuilds);
 
-        //CRM
+        $buildService = new BuildService($store);
+
+        $buildsToProcess = $buildService->skipIntermediateBuilds($result['items']);
 
         $this->logger->addInfo(Lang::get('found_n_builds', count($result['items'])));
 
         $builds = 0;
 
-        while (count($result['items'])) {
-            $build = array_shift($result['items']);
+        while (count($buildsToProcess)) {
+            $build = array_shift($buildsToProcess);
             $build = BuildFactory::getBuild($build);
 
             // Skip build (for now) if there's already a build running in that project:
